requirements.txt
astroid==2.3.3
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
python-dateutil==2.8.1
six==1.14.0
SQLAlchemy==1.3.12
typed-ast==1.4.1
Werkzeug==0.16.0
wrapt==1.11.2
